Remove `IndexingSettings::{sort_field,sort_order}`
