+++ quickwit/quickwit-config/resources/tests/index_config/hdfs-logs.json
@@ -47,8 +47,6 @@
     },
     "indexing_settings": {
         "timestamp_field": "timestamp",
-        "sort_field": "timestamp",
-        "sort_order": "asc",
         "commit_timeout_secs": 61,
         "split_num_docs_target": 10000001,
         "merge_policy": {
