+++ quickwit/quickwit-config/resources/tests/index_config/hdfs-logs.yaml
@@ -32,8 +32,6 @@ retention:
 
 indexing_settings:
   timestamp_field: timestamp
-  sort_field: timestamp
-  sort_order: asc
   commit_timeout_secs: 61
   split_num_docs_target: 10000001
   merge_policy:
