+++ quickwit/quickwit-metastore/test-data/file-backed-index/v0.4-f3d4bd50589e39d506d6754dcfb7453f.json
@@ -85,8 +85,6 @@
         "resources": {
           "heap_size": 3
         },
-        "sort_field": "timestamp",
-        "sort_order": "asc",
         "split_num_docs_target": 10000001,
         "timestamp_field": "timestamp"
       },
