+++ quickwit/quickwit-metastore/test-data/index-metadata/v0.4-fdcafacdb74280ef18f81a45b8dd3273.json
@@ -73,8 +73,6 @@
       "resources": {
         "heap_size": 3
       },
-      "sort_field": "timestamp",
-      "sort_order": "asc",
       "split_num_docs_target": 10000001,
       "timestamp_field": "timestamp"
     },
